* started unlocking station deployment * adjusted vessel mass to take cargo and fitting into account

+++ src/cargo/Item.java
@@ -108,7 +108,7 @@ public class Item implements Serializable {
 
     //Accesors + Mutators
     public double getVolume() {
-        return volume * quantity;
+        return volume * (double)quantity;
     }
 
     public void setVolume(int volume) {
@@ -116,7 +116,7 @@ public class Item implements Serializable {
     }
 
     public double getMass() {
-        return mass * quantity;
+        return mass * (double)quantity;
     }
 
     public void setMass(double mass) {
